requirements.txt
streamlit
pandas
numpy
Pillow
matplotlib